legal-ai-assistant/requirements.txt
# legal-ai-assistant/requirements.txt
google-cloud-aiplatform>=1.38.0
flask>=2.0.1
python-dotenv>=0.19.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
requests>=2.25.1